Added:   - util_trim()   - util_bsearchStringsI()   - Appropriate unit tests
util_bsearchStringsI() searches a table of strings for a particular string
in a case insensitive manner and returns the index of the string within
the table, if found.  This is more direct (easier to call) than the ANSI
bsearch().

Currently, the ASTNode_canonicalize() and supporting functions use it.
UnitKind_forName() should be updated to use it as well.