This file is completely new.  It just happens to have the same name as a file in the old set of XML source files.
Initial import of the new library agnostic XML (LAX) layer.  It presents
a unified (and simplified) SAX-like interface to three popular "backend"
XML libraries: Expat, Xerces-C++, and libXML.  On top of this interface,
it provides an XML stream tokenizer (pull parsing, instead of callback
parsing) and on top of that it provides an XML Document Object Model (DOM).