Fixed a bug which could happen when reading an SBML L2V1 file with an empty math element.
For exmaple, when reading an SBML L2V1 file which contains the following 
three reactions (a kineticlaw in the reaction "r1" contains an empty math 
element which consist of a start and end 'math' tag pair),
the reaction "r2" and "r3" (i.e. all reactions after the reaction "r1") are 
ignored due to the bug and only "r1" is stored in an listOfReactions object.

  <listOfReactions>
    <reaction id="r1" ... >
       ...
       <kineticLaw>
         <math ...>
         </math>
       </kineticLaw>
    </reaction>
    <reaction id="r2" ... >
      ...
    </reaction>
    <reaction id="r3" ... >
      ...
    </reaction>
  </listOfReactions>

An empty math element which consists of single tag ( <math ... /> ) doesn't cause 
the problem.