Removed an invalid test code from TestModel.c.
I noticed that the test code below in test_Model_createKineticLaw_alreadyExists 
has been actually invalid with libSBML 3.x (only valid with 2.x) because the 
behaivour of Model::createKineticLaw() function (internally Reaction::createKineticLaw invoked) 
has been changed in 3.x so as to always delete an existing KineticLaw object and then
recreate a new KineticLaw object.

  -----------------------------------------------------------------------
  r  = Model_createReaction(M);
  kl = Model_createKineticLaw(M);

  fail_unless( Model_createKineticLaw(M) == kl );  <---- **INVALID** (kl may be a junk pointer)
  -----------------------------------------------------------------------

I think the above invalid test code can pass the assertion (in most cases) because 
it is likely that the same address is allocated to the newly created object in the code.
However, the test code fails (in most cases) in language bindings. 
(e.g. always fails in C# binding)