Sorry missed one.
The unit checking code assigns an id to an algebraicRule that allows it to be accessed later. The new internal consistency checking code will complain if an algebraicRule has an id - since it should not have a 'variable' member. Thus the code that uses an id must mark it as internal.