The C implementation of XMLError_getCategoryAsString and XMLError_getSeverityAsString were returning garbage (in MSVC builds). See my message of over a year ago:
This fixes the curious issue of the fact that when
using libsbml compiled with MSVC,
char * return values of some getter functions in the C API  were garbage once returned.

All these functions test correctly with gcc which is why it hasnt been spotted before I tried using them in the MATLAB binding.

I looked into this and found that it is a known pain: see http://www.allegro.cc/forums/thread/592992/0 for a lengthy discussion on the problem.