Changed the error reported when an attribute that used to exist on a component is used in a Level/Version where it is no longer applicable.
When a file with the invalid attribute is read using libSBML a schema error will be reported (this is not a change)

BUT

if a model is created with the invalid attribute and then the consistency is checked an error message stating that the attribute has been removed is reported. (This is a change - previously a schema error was reported)

This applies to 20411; 20615; 21125; 21126; 99206 

NOTE: to be consistent this change needs to apply to 20412 "Celsius has been removed". However the checks on this are more complex and changing it is likely to break code so I havent done it for 3.3.x.