changed the isBooleanFor function to be 'returnsBoolean'. This function no longer takes a model as an argument; if the AST uses a function then it looks for the parent model.
Note this means that if the AST has not yet been added to a document it will not find a model and returnsBoolean will return false (if it uses a function).

I can see a scenario where you might want to know if an AST returns a boolean in the context of a model that it does not yet belong to. I would see this as a different function though.