Problem: Doxygen generates identical HTML for links to code objects (e.g., SBase) as it does for references to documentation sections (via @subpage and the like).  The HTML element it generates in both cases looks like
  <a class="el" ...>...</a>

This makes it impossible to cleanly change the font of code references to be fixed-width typewriter type faces.  They *should* be different, since they are semantically different, and references to code objects should be the same font used for such things as code examples in the documentation.  To solve this problem, I took the approach of wrapping non-code areas (such as lists of @subpage's) with a <div> with a specific CSS class, then in the CSS style file, set the the font of <a> elements within div's of that class.  The <div> code is in the two files 

  doxygen-regular-text-start.html
and
  doxygen-regular-text-end.html

This approach is admittedly suboptimal, because it requires manually adding @htmlinclude for these files in strategic places in the documentation, but it is the best solution I've found so far.

As a consequence, we can now also change the styling of the text in lists of class members and other places to use a fixed-width typewriter style face.  This is set in doxygen-base-stylesheet.css.