Update to fix bugs that show up in validator when running 'checkConsistency' on clearly invalid files.  The problem was that various object-creating functions were not properly guarded against the object creation throwing instead of creating something.  Fixed by doing the try/catch thing that is used elsewhere for UnitDefinitions, and then instead of creating a new Unit, adding it to the UD, then deleting the Unit, just using ud->createUnit(), which doesn't have the possibility of throwing problem, and also is more efficient to boot.
However, a fair amount of things were changed, particularly when UnitDefinitions were being created in several different blocks right next to each other--I changed things where I could so that it was only created once, and then used in the different blocks.  In a few instances, this meant that later logic that used to check (if ud==NULL) had to change to (if ud->getNumUnits()==0), which I believe will work, but should be looked at by someone who is not me.  All the tests do indeed pass (and do test this logic), but it still deserves looking at.