Reverted the changes made to the L3Parser
My reasons:

This is a change in behaviour from previously released behaviour. Rel 5.14.0 had the ability to parse all the new L3V2 function

The only new thing added in this release is the ability to parse '%' as rem.

It will break any code if someone is in the process of implementing support for L3V2 - it breaks the matlab binding support for l3v2 for example.