Add C API to CompModelPlugin from Deviser.  Also needed some non-const versions of C++ functions.
NOTE:  feel free to revert this commit if it is inappropriate for the 'cooling off' period of libsbml; it is not crucial at all, and was actually added as part of my documentation cruise-through.  It can go in the next release if it needs more time to test.  (On the other hand, there's nothing here that's not auto-generated by Deviser, if you consider that reasonably tested.)