Fix flattening of priorities.
This fix probably goes along with a different recent fix where mathml that needed to be converted with a conversion factor where the root ASTNode needed to be converted was not being handled properly.  This means that this test was ensuring that we did things wrong: now that that was fixed, this version is actually correct.