Cleaned up sonarcloud code smells in `SBMLDocument`
Did not clean up all smells since some relied on using C++11 features
(which we want to avoid for compatability). Other remaining smells
include ones which require significant refactoring to fix.

One notable smell was the use of virtual methods in the class
constructors. I've replaced these calls with explicit calls to the
SBMLDocument class (reflecting the fact these would be the methods
called during construction) but I'm unsure if this is the behaviour
intended by the original dev.