Fix the root dirs for importing libsbml source as a module.
I remember this change being made for roadrunner back in the day, but now that I'm working on rr directly, it turns out that the solution didn't quite work, which would have been fantastic information to have had three years ago, but anyway.

The upshot is that the libsbml root dirs need to be set to the *current* directories, not the general directories, or the package files won't work.

There are other issues with using libsbml as a git submodule that are a result of libsbml's CMake system being set up yonks ago, but I think those other issues can be worked around for now.  I've managed to work around most of them so far, at least.