Removed unnecessary copy constructor and destructor from SBMLError
By rule-of-zero, if a class does not require a special copy constructor
or destructor, the compiler-generated versions should be used. In this
case, the copy constructor was performing the same operation the
compiler-defined constructor would have (i.e. calling the base class
constructor) and the virtual destructor was doing nothing. In case
SBMLError is derived from, the implicit destructor will be virtual too
because the base class has a virtual destructor, see [here](
https://isocpp.org/wiki/faq/virtual-functions#virtual-dtors).