Upcast every added XMLError.
Since we return SBMLErrors and not XMLErrors from the error log, we must ensure that all errors are indeed SBMLErrors.

This allows the dynamic casting of all errors from the getError and getErrorWithSeverity functions.

Note that if any XML error had something in the message beyond the basic 'this is what's added for this error code', it will be lost, since there's no way to split that message from the basic message.  If the SBMLError constructor was given the 'getMessage' from the XMLError, the basic message would be duplicated.

Also fix a couple warnings while I'm at it.