Replace unittest.TestCase.assert_ in Python tests
This private method goes away in Python 3.11.

This commit applies a strict mechanical replacement with
unittest.TestCase.assertTrue().

This approach is not ideal—there is often a semantically better public
method, e.g. assertEqual(a, b) instead of assertTrue(a == b)—but it does
ensure we do not accidentally change the *meaning* of a test, especially
distinctions (or lack thereof) between False and None.