Rename 'check' target to be 'use_check'
I'm including libsbml with a bunch of other dependencies for roadrunner and Antimony, and would like to include 'libcheck' in that list.  However, I can't do that because one of libcheck's targets is 'check', which clashes with libsbml's 'check'.

There may be a better thing to rename the target to, and there may be other issues in testing, like if anything tries to run 'make check'; I'm happy to adjust if need be (or just leave this as a branch).