Only call loadPlugins once in constructors.
libsbml is leaking memory in the loadPlugins functions, and my working theory is that this is because some constructors call it multiple times: once in the base class, then again in the derived class.  This change removes *all* calls of loadPlugins for all derived classes where the base class also calls loadPlugins.

I'm not sure why we don't call loadPlugins in the SBase constructor directly, but changing that would be a fairly significant design change, so no need to try it now.

I believe that we cannot do the same thing with 'connectToChild', because it becomes important when it's the derived class that overrides the base definition for it, and I *think* that a base class constructor would not call the derived class version.  If I'm wrong, we can revisit, but it's also true that connectToChild doesn't leak memory, so it's not dangerous to call it multiple times.

'setElementNamespace' also only needs to be called once, but this is sort of incidental to the other cleanups.

I have not yet tested whether this change actually fixes the leak--needed to check this in from my development machine so I could check it out on my valgrind-enabled machine.