Encode all &, <, and > in markdown.
This is non-standard markdown!  But it has the advantage of creating valid HTML more often, so I think on balance it's a win?

This means that this translator will *not* translate valid HTML in markdown to actual HTML.  I figured this was a more uncommon occurrence than (say) all the biomodels notes that have "<=>" as a reaction symbol, or that have '&' in a URL.

We can drop this (I think) if https://github.com/tim-gromeyer/html2md/issues/158 is addressed.  But Maddy might need updating as well.

Anyway: hack for now, but happy to revert if people think the cost is too severe.